+++ lib/service/computetask_test.go
@@ -285,10 +285,6 @@ func TestRegisterTrainTask(t *testing.T) {
 		AssetKey:  newTrainTask.Key,
 		EventKind: asset.EventKind_EVENT_ASSET_CREATED,
 		Asset:     &asset.Event_ComputeTask{ComputeTask: storedTask},
-		Metadata: map[string]string{
-			"status": storedTask.Status.String(),
-			"worker": dataManager.Owner,
-		},
 	}
 	es.On("RegisterEvents", expectedEvent).Once().Return(nil)
 
@@ -472,10 +468,6 @@ func TestRegisterCompositeTaskWithCompositeParents(t *testing.T) {
 		AssetKey:  newTask.Key,
 		EventKind: asset.EventKind_EVENT_ASSET_CREATED,
 		Asset:     &asset.Event_ComputeTask{ComputeTask: storedTask},
-		Metadata: map[string]string{
-			"status": storedTask.Status.String(),
-			"worker": dataManager.Owner,
-		},
 	}
 	es.On("RegisterEvents", expectedEvent).Once().Return(nil)