+++ lib/service/computetaskstate.go
@@ -249,10 +249,7 @@ func (s *ComputeTaskService) onStateChange(e *fsm.Event) {
 		AssetKind: asset.AssetKind_ASSET_COMPUTE_TASK,
 		Asset:     &asset.Event_ComputeTask{ComputeTask: task},
 		Metadata: map[string]string{
-			"status":           task.Status.String(),
-			"reason":           reason,
-			"worker":           task.Worker,
-			"compute_plan_key": task.ComputePlanKey,
+			"reason": reason,
 		},
 	}
 	err = s.GetEventService().RegisterEvents(event)