+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- Metadata set in events (#787)
+
 ## [0.17.0] - 2022-06-20
 
 ### Added